requirements.txt
requests
python-dotenv
pyyaml
pandas
hsfs